requirements.txt
Flask>=2.3.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
gunicorn>=20.0.0
setuptools>=68.0.0
wheel>=0.38.0
